feat: expose external types
